Revert "separate out helpers"
This reverts commit 47c775840334df9e289de3f6ce71fb1ceee03606.
